--- a/Cargo.toml
+++ b/Cargo.toml
@@ -30,7 +30,5 @@ conv     = { version = "0.3", default-features = false }
 compiletest_rs = { version = "0.1.1", optional = true }
 
 [features]
-default       = ["verbose_error"]
-verbose_error = []
 # Feature for running extra (compiletime fail) tests on nightly
 unstable      = ["compiletest_rs"]