--- a/src/parse_result.rs
+++ b/src/parse_result.rs
@@ -354,7 +354,7 @@ impl<'a, I, T: fmt::Debug, E> ParseResult<'a, I, T, E> {
     ///
     /// let r = token(Input::new(b"a"), b'b');
     ///
-    /// assert_eq!(r.unwrap_err(), Error::Expected(98));
+    /// assert_eq!(r.unwrap_err(), Error::expected(98));
     /// ```
     ///
     /// ```{.should_panic}
@@ -363,7 +363,7 @@ impl<'a, I, T: fmt::Debug, E> ParseResult<'a, I, T, E> {
     /// let r = token(Input::new(b"a"), b'a');
     ///
     /// // Panics with "called `ParseResult::unwrap_err` on a success state: 97"
-    /// assert_eq!(r.unwrap_err(), Error::Expected(98));
+    /// assert_eq!(r.unwrap_err(), Error::expected(98));
     /// ```
     #[inline]
     pub fn unwrap_err(self) -> E {