@m4rw3r 2016-02-28
Does still not match the performance of the noop error, it will probably be worth keeping as a feature for really performance-intensive stuff (and they usually are fine with an error offset only I believe).