@m4rw3r 2016-03-16
The `compiletest_rs`  crate is broken on current nightly.